WIP for comment email notifications
