+++ LB.PhotoGalleries.Application/Servers/ImageServer.cs
