+++ LB.PhotoGalleries.Worker/Properties/launchSettings.json
@@ -1,7 +1,8 @@
 {
   "profiles": {
     "LB.PhotoGalleries.Worker": {
-      "commandName": "Project"
+      "commandName": "Project",
+      "commandLineArgs": "\"C:\\Users\\JayVanderZant\\OneDrive - Tetron\\LB\\LB Photos\\Worker\\appsettings.Development.json\""
     }
   }
 }
